requirements.txt
#  Database
psycopg2-binary==2.9.5

#  Enviroment variables
python-dotenv==0.21.0

pyserial==3.5
